+++ src/YINI.ts
@@ -1,9 +1,13 @@
 import { isDebug, isDev } from './config/env'
-import { ErrorDataHandler } from './core/ErrorDataHandler'
+import { ErrorDataHandler } from './core/errorDataHandler'
 import { isOptionsObjectForm } from './core/options/normalizeOptions'
 import { YiniRuntime } from './core/runtime'
-import { IAllUserOptions, TJSObject, TPreferredFailLevel } from './core/types'
-import { _parseMain } from './parseEntry'
+import {
+    IAllUserOptions,
+    ParsedObject,
+    TPreferredFailLevel,
+    YiniParseResult,
+} from './types'
 import { debugPrint, devPrint, printObject } from './utils/print'
 
 const DEFAULT_TAB_SIZE = 4 // De facto "modern default" (even though traditionally/historically it's 8).
@@ -61,7 +65,7 @@ export default class YINI {
      *   - `'ignore-errors'`       : Continue parsing despite errors; log them and attempt recovery.
      *   - `'errors'`              : Stop parsing on the first error.
      *   - `'warnings-and-errors'` : Stop parsing on the first warning **or** error.
-     * @param includeMetaData    If `true`, return additional metadata (e.g. warnings, statistics) alongside the parsed object.
+     * @param includeMetadata    If `true`, return additional metadata (e.g. warnings, statistics) alongside the parsed object.
      *
      * @returns A JavaScript object representing the parsed YINI content.
      */
@@ -73,14 +77,15 @@ export default class YINI {
         yiniContent: string,
         strictMode?: boolean,
         failLevel?: TPreferredFailLevel,
-        includeMetaData?: boolean,
-    ): TJSObject
+        includeMetadata?: boolean,
+    ): ParsedObject | YiniParseResult
 
     /**
      * Parse inline YINI content into a JavaScript object, using an options object for configuration.
      *
      * @param yiniContent      YINI code as a string (multi‑line content supported).
      * @param options Optional settings to customize parsing and/or results, useful if you need more control.
+     *        For all options, see types/IAllUserOptions.
      *
      * @param options.failLevel - Minimum severity that should cause the parse to fail.
      *   Accepts:
@@ -89,15 +94,15 @@ export default class YINI {
      *     `'warnings-and-errors'` - Stop parsing on the first warning or error.
      *   (Type: TPreferredFailLevel; exact behavior is implementation-defined.)
      * @param options.includeDiagnostics - Include diagnostics in the returned metadata.
-     *   Requires: `includeMetaData = true`. Ignored otherwise.
-     * @param options.includeMetaData - Attach a metadata object to the parse result
+     *   Requires: `includeMetadata = true`. Ignored otherwise.
+     * @param options.includeMetadata - Attach a metadata object to the parse result
      *   (e.g., timings, diagnostics).
      * @param options.includeTiming - Include timing information for parser phases in metadata.
-     *   Requires: `includeMetaData = true`. Ignored otherwise.
+     *   Requires: `includeMetadata = true`. Ignored otherwise.
      * @param options.onDuplicateKey - Strategy/handler when encountering a duplicate key.
      *   Allowed values: `'warn-and-keep-first'` | `'warn-and-overwrite'` | `'keep-first'` (silent, first wins) | `'overwrite'` (silent, last wins) | `'error'`.
      * @param options.preserveUndefinedInMeta - Keep properties with value `undefined` inside
-     *   the returned metadata. Requires: `includeMetaData = true`. Ignored otherwise.
+     *   the returned metadata. Requires: `includeMetadata = true`. Ignored otherwise.
      * @param options.requireDocTerminator - Controls whether a document terminator is required.
      *   Allowed values: `'optional'` | `'warn-if-missing'` | `'required'`.
      * @param options.strictMode - Enable stricter syntax and well-formedness checks according
@@ -116,7 +121,7 @@ export default class YINI {
     public static parse(
         yiniContent: string,
         options?: IAllUserOptions,
-    ): TJSObject
+    ): ParsedObject | YiniParseResult
 
     // --- Single implementation --------------------------------------------
     // Implementation method (not declared with arrow function) for both method overload signatures.
@@ -125,14 +130,14 @@ export default class YINI {
         yiniContent: string,
         arg2?: boolean | IAllUserOptions, // strictMode | options
         failLevel: TPreferredFailLevel = 'auto',
-        includeMetaData = false,
-    ): TJSObject {
+        includeMetadata = false,
+    ): ParsedObject | YiniParseResult {
         debugPrint('-> Entered static parse(..) in class YINI\n')
 
         // // Runtime guard to catch illegal/ambiguous calls coming from JS or any-cast code
         // if (
         //     isOptionsObjectForm(arg2) &&
-        //     (failLevel !== 'auto' || includeMetaData !== false)
+        //     (failLevel !== 'auto' || includeMetadata !== false)
         // ) {
         //     throw new TypeError(
         //         'Invalid call: when providing an options object, do not also pass positional parameters.',
@@ -158,11 +163,11 @@ export default class YINI {
         //         strictMode:
         //             (arg2 as boolean | undefined) ?? defaultOptions.strictMode,
         //         failLevel,
-        //         includeMetaData,
+        //         includeMetadata,
         //     }
         // }
 
-        // if (userOpts.includeMetaData && _runtimeInfo.sourceType === 'Inline') {
+        // if (userOpts.includeMetadata && _runtimeInfo.sourceType === 'Inline') {
         //     const lineCount = yiniContent.split(/\r?\n/).length // Counts the lines.
         //     const sha256 = computeSha256(yiniContent) // NOTE: Compute BEFORE any possible tampering of content.
 
@@ -203,7 +208,7 @@ export default class YINI {
                   yiniContent,
                   arg2 as boolean | undefined,
                   failLevel,
-                  includeMetaData,
+                  includeMetadata,
               )
         debugPrint('==== End call parse ==========================\n')
 
@@ -229,7 +234,7 @@ export default class YINI {
      *   - `'ignore-errors'`       : Continue parsing despite errors; log them and attempt recovery.
      *   - `'errors'`              : Stop parsing on the first error.
      *   - `'warnings-and-errors'` : Stop parsing on the first warning **or** error.
-     * @param includeMetaData    If `true`, return additional metadata (e.g. warnings, statistics) alongside the parsed object.
+     * @param includeMetadata    If `true`, return additional metadata (e.g. warnings, statistics) alongside the parsed object.
      *
      * @returns A JavaScript object representing the parsed YINI content.
      */
@@ -241,14 +246,15 @@ export default class YINI {
         filePath: string,
         strictMode?: boolean,
         failLevel?: TPreferredFailLevel,
-        includeMetaData?: boolean,
-    ): TJSObject
+        includeMetadata?: boolean,
+    ): ParsedObject | YiniParseResult
 
     /**
      * Parse a YINI file into a JavaScript object, using an options object for configuration.
      *
      * @param yiniFile Path to the YINI file.
      * @param options Optional settings to customize parsing and/or results, useful if you need more control.
+     *        For all options, see types/IAllUserOptions.
      *
      * @param options.failLevel - Minimum severity that should cause the parse to fail.
      *   Accepts:
@@ -257,15 +263,15 @@ export default class YINI {
      *     `'warnings-and-errors'` - Stop parsing on the first warning or error.
      *   (Type: TPreferredFailLevel; exact behavior is implementation-defined.)
      * @param options.includeDiagnostics - Include diagnostics in the returned metadata.
-     *   Requires: `includeMetaData = true`. Ignored otherwise.
-     * @param options.includeMetaData - Attach a metadata object to the parse result
+     *   Requires: `includeMetadata = true`. Ignored otherwise.
+     * @param options.includeMetadata - Attach a metadata object to the parse result
      *   (e.g., timings, diagnostics).
      * @param options.includeTiming - Include timing information for parser phases in metadata.
-     *   Requires: `includeMetaData = true`. Ignored otherwise.
+     *   Requires: `includeMetadata = true`. Ignored otherwise.
      * @param options.onDuplicateKey - Strategy/handler when encountering a duplicate key.
      *   Allowed values: `'warn-and-keep-first'` | `'warn-and-overwrite'` | `'keep-first'` (silent, first wins) | `'overwrite'` (silent, last wins) | `'error'`.
      * @param options.preserveUndefinedInMeta - Keep properties with value `undefined` inside
-     *   the returned metadata. Requires: `includeMetaData = true`. Ignored otherwise.
+     *   the returned metadata. Requires: `includeMetadata = true`. Ignored otherwise.
      * @param options.requireDocTerminator - Controls whether a document terminator is required.
      *   Allowed values: `'optional'` | `'warn-if-missing'` | `'required'`.
      * @param options.strictMode - Enable stricter syntax and well-formedness checks according
@@ -284,7 +290,7 @@ export default class YINI {
     public static parseFile(
         filePath: string,
         options?: IAllUserOptions,
-    ): TJSObject
+    ): ParsedObject | YiniParseResult
 
     // --- Single implementation --------------------------------------------
     // Implementation method (not declared with arrow function) for both method overload signatures.
@@ -293,15 +299,15 @@ export default class YINI {
         filePath: string,
         arg2?: boolean | IAllUserOptions, // strictMode | options
         failLevel: TPreferredFailLevel = 'auto',
-        includeMetaData = false,
-    ): TJSObject {
+        includeMetadata = false,
+    ): ParsedObject | YiniParseResult {
         debugPrint('-> Entered static parseFile(..) in class YINI\n')
         debugPrint('Current directory = ' + process.cwd())
 
         // // Runtime guard to catch illegal/ambiguous calls coming from JS or any-cast code
         // if (
         //     isOptionsObjectForm(arg2) &&
-        //     (failLevel !== 'auto' || includeMetaData !== false)
+        //     (failLevel !== 'auto' || includeMetadata !== false)
         // ) {
         //     throw new TypeError(
         //         'Invalid call: when providing an options object, do not also pass positional parameters.',
@@ -328,7 +334,7 @@ export default class YINI {
         //         strictMode:
         //             (arg2 as boolean | undefined) ?? defaultOptions.strictMode,
         //         failLevel,
-        //         includeMetaData,
+        //         includeMetadata,
         //     }
         // }
 
@@ -354,7 +360,7 @@ export default class YINI {
         // _runtimeInfo.sourceType = 'File'
         // _runtimeInfo.fileName = filePath
 
-        // if (userOpts.includeMetaData) {
+        // if (userOpts.includeMetadata) {
         //     _runtimeInfo.lineCount = content.split(/\r?\n/).length // Counts the lines.
         //     _runtimeInfo.fileByteSize = fileByteSize
         //     _runtimeInfo.timeIoMs = +(timeEndMs - timeStartMs).toFixed(3)
@@ -382,7 +388,7 @@ export default class YINI {
                   filePath,
                   arg2 as boolean | undefined,
                   failLevel,
-                  includeMetaData,
+                  includeMetadata,
               )
 
         debugPrint('==== End call parse ==========================\n')