+++ src/core/options/normalizeOptions.ts
@@ -1,10 +1,6 @@
 import { isDebug, isDev } from '../../config/env'
-import {
-    IAllUserOptions,
-    IParseCoreOptions,
-    TBailSensitivityLevel,
-    TParserMode,
-} from '../types'
+import { IAllUserOptions, TBailSensitivityLevel } from '../../types'
+import { IParseCoreOptions, TParserMode } from '../internalTypes'
 
 export const toCoreOptions = (
     bailLevel: TBailSensitivityLevel,
@@ -13,7 +9,7 @@ export const toCoreOptions = (
     return {
         isStrict: userOpts.strictMode,
         bailSensitivity: bailLevel,
-        isIncludeMeta: userOpts.includeMetaData,
+        isIncludeMeta: userOpts.includeMetadata,
         isWithDiagnostics: isDev() || isDebug() || userOpts.includeDiagnostics,
         isWithTiming: isDev() || isDebug() || userOpts.includeTiming,
         isKeepUndefinedInMeta: isDebug() || userOpts.preserveUndefinedInMeta,
@@ -33,7 +29,7 @@ export const isOptionsObjectForm = (v: unknown): v is IAllUserOptions => {
         // but this keeps accidental booleans/strings out.
         ('strictMode' in (v as any) ||
             'failLevel' in (v as any) ||
-            'includeMetaData' in (v as any) ||
+            'includeMetadata' in (v as any) ||
             'includeDiagnostics' in (v as any) ||
             'includeTiming' in (v as any) ||
             'preserveUndefinedInMeta' in (v as any) ||