+++ src/core/options/parserOptionsConstants.ts
@@ -1,4 +1,5 @@
-import { IAllUserOptions, TParserMode } from '../types'
+import { IAllUserOptions } from '../../types'
+import { TParserMode } from '../internalTypes'
 
 export const getDefaultOptions = (mode: TParserMode) =>
     mode === 'strict' ? DEFAULT_STRICT_OPTS : DEFAULT_LENIENT_OPTS
@@ -8,7 +9,7 @@ export type NormalizedOptions = Required<
         IAllUserOptions,
         | 'strictMode'
         | 'failLevel'
-        | 'includeMetaData'
+        | 'includeMetadata'
         | 'includeDiagnostics'
         | 'includeTiming'
         | 'preserveUndefinedInMeta'
@@ -23,7 +24,7 @@ export type NormalizedOptions = Required<
 export const BASE_DEFAULTS: NormalizedOptions = {
     strictMode: false,
     failLevel: 'auto',
-    includeMetaData: false,
+    includeMetadata: false,
     includeDiagnostics: false,
     includeTiming: false,
     preserveUndefinedInMeta: false,